fix: remove context exports and refactor context file: collapsible
