--- packages/kit-headless/src/components/collapsible/collapsible-context-id.ts
@@ -1,4 +1,0 @@
-import { createContextId } from '@builder.io/qwik';
-import { type CollapsibleContext } from './collapsible-context.type';
-
-export const collapsibleContextId = createContextId<CollapsibleContext>('collapsible');
