+++ packages/kit-headless/src/components/collapsible/index.ts
@@ -1,5 +1,3 @@
 export { HCollapsible as Root } from './collapsible';
 export { HCollapsibleContent as Content } from './collapsible-content';
 export { HCollapsibleTrigger as Trigger } from './collapsible-trigger';
-export * from './collapsible-context.type';
-export * from './collapsible-context-id';
